team scoped /api/billings
