+++ pages/api/teams/[teamId]/billing/manage.ts
@@ -1,0 +1,51 @@
+import { NextApiRequest, NextApiResponse } from "next";
+import { getServerSession } from "next-auth/next";
+import { authOptions } from "../../../auth/[...nextauth]";
+import { CustomUser } from "@/lib/types";
+import { stripe } from "@/lib/stripe";
+import prisma from "@/lib/prisma";
+import { errorHanlder } from "@/lib/errorHandler";
+
+export default async function handle(
+  req: NextApiRequest,
+  res: NextApiResponse
+) {
+  if (req.method === "POST") {
+    // POST /api/billing/manage – manage a user's subscription
+    const session = await getServerSession(req, res, authOptions);
+    if (!session) {
+      res.status(401).end("Unauthorized");
+      return;
+    }
+
+    const { teamId } = req.query as { teamId: string };
+    try {
+      const team = await prisma.team.findUnique({
+        where: {
+          id: teamId,
+        },
+        select: {
+          stripeId: true,
+        },
+      });
+
+      if (!team) {
+        return res.status(400).json({ error: "Team does not exists" });
+      }
+      if (!team.stripeId) {
+        return res.status(400).json({ error: "No Stripe customer ID" });
+      }
+
+      const { url } = await stripe.billingPortal.sessions.create({
+        customer: team.stripeId,
+        return_url: `${process.env.NEXTAUTH_URL}/settings/billing`,
+      });
+      return res.status(200).json(url);
+    } catch (error) {
+      errorHanlder(error, res);
+    }
+  } else {
+    res.setHeader("Allow", ["POST"]);
+    return res.status(405).end(`Method ${req.method} Not Allowed`);
+  }
+}
