+++ pages/api/teams/[teamId]/billing/upgrade.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { getServerSession } from "next-auth/next";
-import { authOptions } from "../auth/[...nextauth]";
+import { authOptions } from "../../../auth/[...nextauth]";
 import { CustomUser } from "@/lib/types";
 import { stripe } from "@/lib/stripe";
 
@@ -9,17 +9,22 @@ export default async function handle(
   res: NextApiResponse
 ) {
   if (req.method === "POST") {
-    // POST /api/billing/upgrade
+    // POST /api/teams/:teamId/billing/upgrade
     const session = await getServerSession(req, res, authOptions);
     if (!session) {
       res.status(401).end("Unauthorized");
       return;
     }
 
-    const { priceId } = req.query as { priceId: string };
+    const { teamId, priceId } = req.query as {
+      teamId: string;
+      priceId: string;
+    };
+
+    const userEmail = (session.user as CustomUser).email;
 
     const stripeSession = await stripe.checkout.sessions.create({
-      customer_email: (session.user as CustomUser).email ?? undefined,
+      customer_email: userEmail ?? undefined,
       billing_address_collection: "required",
       success_url: `${process.env.NEXTAUTH_URL}/settings/billing?success=true`,
       cancel_url: `${process.env.NEXTAUTH_URL}/settings/billing`,
@@ -32,7 +37,7 @@ export default async function handle(
       },
       mode: "subscription",
       allow_promotion_codes: true,
-      client_reference_id: (session.user as CustomUser).id,
+      client_reference_id: teamId,
     });
     return res.status(200).json(stripeSession);
   } else {
